pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "openclaw-voice-server"
version = "0.0.4"
description = "Standalone browser-to-OpenClaw voice server."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.11.0",
  "httpx>=0.28.0",
  "numpy>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]
stt-faster-whisper = [
  "faster-whisper>=1.1.0",
]
stt-whisper = [
  "openai-whisper>=20240930",
]
tts-edge = [
  "edge-tts>=6.1.0",
]

[project.scripts]
openclaw-voice-server = "openclaw_voice_server.app:main"
openclaw-calibrate-command = "openclaw_voice_server.calibrate_commands:main"

[tool.hatch.build.targets.wheel]
packages = ["src/openclaw_voice_server"]

[tool.hatch.build.targets.wheel.force-include]
"src/openclaw_voice_server/static" = "openclaw_voice_server/static"

[tool.hatch.build.targets.sdist]
include = [
  "/src/openclaw_voice_server/**/*.py",
  "/src/openclaw_voice_server/static/**/*",
  "/tests/**/*",
  "/README.md",
  "/pyproject.toml",
  "/.env.example",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
